+++ Makefile
@@ -45,6 +45,7 @@ PYFILES := dml/__init__.py			\
           dml/slotsmeta.py			\
           dml/structure.py			\
           dml/symtab.py				\
+          dml/warnings.py			\
 	  __main__.py
 
 PYUNIT_FILES := $(wildcard $(DMLC_DIR)/py/dml/*_test.py)
